% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_overview.R
\name{country_status}
\alias{country_status}
\title{compile an overview report of each countries status}
\usage{
country_status(rcm = NULL, subs = NULL, update_rcm = T)
}
\arguments{
\item{rcm}{can be ommited}

\item{subs}{can be ommited}
}
\value{
nothing interesting, but writes a html file into the working directory and opens it in the browser
}
\description{
compile an overview report of each countries status
}
