% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_overview.R
\name{country_status_markdown}
\alias{country_status_markdown}
\title{Map results to an output template}
\usage{
country_status_markdown(rcm, delayed, inconsistencies, path = "./",
  filename = "country_overview.html")
}
\description{
Map results to an output template
}
