% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{date_arrived}
\alias{date_arrived}
\title{is a date today or older?}
\usage{
date_arrived(date)
}
\arguments{
\item{date}{date(s) to check}
}
\value{
logical vector if date is today or in the passed
}
\description{
is a date today or older?
}
