% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_allocation.R
\name{filter_validated_in_month}
\alias{filter_validated_in_month}
\title{filter rows from rcm that have state 'validated' and date in a certain month / year}
\usage{
filter_validated_in_month(rcm, month, year = 2019)
}
\arguments{
\item{rcm}{the RCM}

\item{month}{the month to keep (single integer)}

\item{year}{the year to keep (single integer; defaults to 2019)}
}
\description{
filter rows from rcm that have state 'validated' and date in a certain month / year
}
