% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_api.R
\name{g_sheets_append_row}
\alias{g_sheets_append_row}
\title{wrapper for google sheets api to append a new row}
\usage{
g_sheets_append_row(value,
  spreadsheetId = "1iNt__-uMMBTbLEsJkiIXglPJ4GK-9UCVqC7awhMTXF8", ...)
}
\arguments{
\item{value}{a character vector of values to put in the row}
}
\description{
wrapper for google sheets api to append a new row
}
