% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_api.R
\name{g_sheets_put}
\alias{g_sheets_put}
\title{generally we're using googlesheets package for authentification, but then httr package to call the google drive directly.
the workhorse is g_sheets_put which is used in most other functions that edit the remote RCM.}
\usage{
g_sheets_put(row, col, value,
  spreadsheetId = "1Quu2P6z-uA2H64eQENWJkNIOGIfnsdXgKNg4qdiCvXc", ...)
}
\arguments{
\item{row}{the number of the row we want to fill}

\item{col}{the column to fill as a letter character, excel style: ( "A", "B", .... , "Z", "BA", ..., "ZZ")}

\item{value}{the value to put into the cell}

\item{spreadsheetId}{we could use this for other spreadsheets, but it's not working at the moment. Throwing an error if this param is changed}
}
\description{
wrapper for google sheets api to change single value
}
