% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{g_sheets_update_index}
\alias{g_sheets_update_index}
\title{fills a column on google drive with incremental indices to ensure row matching works as expected}
\usage{
g_sheets_update_index(col = "AR",
  spreadsheetId = "1wX5k3cETrCbnw4vpfY07eSzTyWX6AwmJmxJQwPahrSk", ...)
}
\description{
fills a column on google drive with incremental indices to ensure row matching works as expected
}
