% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{get_gdrive_row}
\alias{get_gdrive_row}
\title{which row on google drive matches a file.id?}
\usage{
get_gdrive_row(file.id, rcm = NULL)
}
\arguments{
\item{file.id}{the file id name of the item}

\item{rcm}{you can provide a custom rcm to use for row identification.
Not recommended (except for debugging purposes)
this will return the wrong row if the this is not a raw rcm (from rcm_download(raw=T)) or if the RCM online changed since it was downloaded}
}
\value{
(estimated) google drive row index of file.id
}
\description{
which row on google drive matches a file.id?
}
