% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_dataunit_focalpoints.R
\name{hq_focal_point}
\alias{hq_focal_point}
\title{who is the focal point for the rcid?}
\usage{
hq_focal_point(rcid)
}
\value{
a vector of surnames of the HQ data unit focal point for the research cycle
}
\description{
who is the focal point for the rcid?
}
