% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_set_submission_date.R
\name{is_date_format_in_rcm_style}
\alias{is_date_format_in_rcm_style}
\title{check if a date string is in the format that matches the research cycle matrix standard for dates}
\usage{
is_date_format_in_rcm_style(date_string)
}
\description{
check if a date string is in the format that matches the research cycle matrix standard for dates
}
