% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_shiny.R
\name{rcm_add_validation_button}
\alias{rcm_add_validation_button}
\title{add validation button html to rcm}
\usage{
rcm_add_validation_button(rcm)
}
\value{
rcm with button link
}
\description{
add validation button html to rcm
}
