% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{rcm_add_value}
\alias{rcm_add_value}
\title{add a value on google drive based on the file.id}
\usage{
rcm_add_value(file.id, column, value)
}
\description{
add a value on google drive based on the file.id
}
