% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_allocation.R
\name{rcm_allocation_table}
\alias{rcm_allocation_table}
\title{create allocation table}
\usage{
rcm_allocation_table(who = ".", rcm = NULL, rcm_raw = NULL,
  include.file.ids = c(), month, year = 2019)
}
\arguments{
\item{who}{regex pattern matching HQ Focal Point (defaults to  "." which matches all)}

\item{rcm}{the research cycle matrix. leave empty to download fresh}

\item{rcm_raw}{the raw research cycle matrix; get with rcm_download(raw = T) or levae empty to download fresh}

\item{include.file.ids}{vector of file ids that should be included even if they do not match the focal point pattern in "who" parameter}

\item{month}{integer of the month for which to create the allocation table}

\item{year}{integer of the year for which to create the allocation table (defaults to 2019; that default will likely change, so even if it's 2019, you should provide it in case that stable future behaviour is needed)}
}
\description{
create allocation table
}
