% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_api.R
\name{rcm_change_value}
\alias{rcm_change_value}
\title{change a value on google drive based on the file.id}
\usage{
rcm_change_value(file.id, column, value)
}
\arguments{
\item{file.id}{as a string}

\item{column}{best to just put the google drive colum letters (like "A" for the first column, or "AF")}

\item{value}{the value to put in the column, probably safest to give a character string}
}
\description{
change a value on google drive based on the file.id
}
