% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_inconsistencies.R
\name{rcm_check}
\alias{rcm_check}
\title{Check inconsistencies in RCM}
\usage{
rcm_check(rcm, check.archived = F, check.validated = F)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
data frame with list of inconsistencies
}
\description{
Check inconsistencies in RCM
}
