% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_check_csv.R
\name{rcm_check_csv}
\alias{rcm_check_csv}
\title{write matrix issues to file for each HQ focal point}
\usage{
rcm_check_csv(rcm = NULL)
}
\arguments{
\item{rcm}{use rcm_download(gdrive_links=F). can be ommited (then downloading from gdrive)}
}
\value{
writes csv files to current wd, and also opens them in excel directly!
}
\description{
write matrix issues to file for each HQ focal point
}
