% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_overview.R
\name{rcm_dashboard}
\alias{rcm_dashboard}
\title{compile an overview report of each countries status}
\usage{
rcm_dashboard(rcm = NULL, subs = NULL, update_rcm = T)
}
\arguments{
\item{rcm}{can be ommited}

\item{subs}{can be ommited}
}
\value{
nothing interesting, but writes a html file into the working directory and opens it in the browser
}
\description{
compile an overview report of each countries status
}
