% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_days_with_hq}
\alias{rcm_days_with_hq}
\title{how long has the item been with HQ?}
\usage{
rcm_days_with_hq(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
Date time difference vector
}
\description{
how long has the item been with HQ?
}
