% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_deadline_status}
\alias{rcm_deadline_status}
\title{Days until a given date as ordinal categories}
\usage{
rcm_deadline_status(date)
}
\arguments{
\item{date}{a vector of dates}
}
\value{
vector with ordinal values "more than 30 days", "within 30 days", "within 14 days", "within 7 days", "tomorrow", "today" and "overdue"
}
\description{
Changes a date of a deadline into words such as "today", "tomorrow", "within 30 days"
}
