% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo_delayed.R
\name{rcm_delayed_to_google_sheet}
\alias{rcm_delayed_to_google_sheet}
\title{write delays to google sheet}
\usage{
rcm_delayed_to_google_sheet(rcm)
}
\arguments{
\item{todo}{todo list (see ?todo_download)}
}
\description{
write delays to google sheet
}
\seealso{
todo_delayed_browse()
}
