% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_preprocessing.R
\name{rcm_fill_dates}
\alias{rcm_fill_dates}
\title{filling dates (used in rcm_download when fill_dates=T)}
\usage{
rcm_fill_dates(rcm)
}
\description{
filling dates (used in rcm_download when fill_dates=T)
}
