% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_inconsistencies.R
\name{rcm_find_issue}
\alias{rcm_find_issue}
\title{applies an issue checking function to the RCM}
\usage{
rcm_find_issue(rcm, get_issue_ids_fun, issue_name)
}
\description{
applies an issue checking function to the RCM
}
