% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_search.R
\name{rcm_find_row_by_file.id}
\alias{rcm_find_row_by_file.id}
\title{Search RCM for terms and return the row}
\usage{
rcm_find_row_by_file.id(rcm, search, unit = NULL)
}
\arguments{
\item{search}{single string with search terms separated by a simple space. Search is not case sensitive. The best match will be returned. If there are multiple matches with the best match score, the user is prompted to select one.}
}
\value{
the RCM row as a data frame with a single row
}
\description{
Search RCM for terms and return the row
}
