% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_plots.R
\name{rcm_gant}
\alias{rcm_gant}
\title{Gant chart of all RCM items}
\usage{
rcm_gant(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\description{
Gant chart of all RCM items
}
