% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_has_identified_status}
\alias{rcm_has_identified_status}
\title{is the RCM status coercible into a standard state?}
\usage{
rcm_has_identified_status(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
is the RCM status coercible into a standard state?
}
