% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_is_data_unit_item}
\alias{rcm_is_data_unit_item}
\title{do RCM rows belong to the data unit?}
\usage{
rcm_is_data_unit_item(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
do RCM rows belong to the data unit?
}
