% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_inconsistencies.R
\name{rcm_missing_data_unit_items}
\alias{rcm_missing_data_unit_items}
\title{Identify research cycles that do not have data and analysis items}
\usage{
rcm_missing_data_unit_items(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
data frame with columns..
- data.missing ("exists" or "missing")
- analysis.missing ("exists" or "missing")
- RCID (the research cycle ID)
}
\description{
Identify research cycles that do not have data and analysis items
}
