% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_passed_milestone}
\alias{rcm_passed_milestone}
\title{has an item missed its milestone?}
\usage{
rcm_passed_milestone(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
has an item missed its milestone?
}
