% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_past_planned_submission}
\alias{rcm_past_planned_submission}
\title{has the submission date passed?}
\usage{
rcm_past_planned_submission(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
has the submission date passed?
}
