% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_prefill_research_tracker.R
\name{rcm_prefill_research_tracker}
\alias{rcm_prefill_research_tracker}
\title{create a prefilled monthly RC tracker xlsx file}
\usage{
rcm_prefill_research_tracker(output_xlsx_file_name = paste0("tracker_",
  Sys.Date()), rcm = NULL)
}
\arguments{
\item{output_xlsx_file_name}{the path to the newly created file}

\item{the}{rcm; see rcm_download()}
}
\value{
writes an xlsx template with prefilled file id rows as a side effect (if xlsx package is installed). Returns a data.frame with the rows that were prefilled
}
\description{
create a prefilled monthly RC tracker xlsx file
}
\details{
only creates an xslx template file if the _xlsx package is installed_.
In any case this function will return (invisibly) a data.frame with the rows that need to be put in the tracker.
if you want an xlsx template file to be created, you need to install.packages("xlsx"). Careful you must have compatible R and Java versions for this to work (both 32 or both 64 bit)
}
