% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_allocation.R
\name{rcm_set_hours_worked}
\alias{rcm_set_hours_worked}
\title{save the hours worked in the rcm}
\usage{
rcm_set_hours_worked(file.id, hours)
}
\arguments{
\item{file.id}{the file id}

\item{hours}{the number of hours worked on the file id}
}
\value{
the file id
}
\description{
save the hours worked in the rcm
}
