% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{rcm_set_HQ_focal_point}
\alias{rcm_set_HQ_focal_point}
\title{set the name of the person handling the item at HQ in the gdrive RCM}
\usage{
rcm_set_HQ_focal_point(file.id, name = NULL)
}
\arguments{
\item{file.id}{the file id of the item to change}

\item{name}{the name of the person as a character string.}
}
\value{
returns nothing important
}
\description{
set the name of the person handling the item at HQ in the gdrive RCM
}
