% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_set_submission_date.R
\name{rcm_set_submission_date}
\alias{rcm_set_submission_date}
\title{set the submission date in the gdrive RCM}
\usage{
rcm_set_submission_date(file.id, date)
}
\arguments{
\item{file.id}{the file id of the item to change}

\item{date}{the date to set it to. Must be of class "POSIXct".}
}
\value{
returns nothing important
}
\description{
set the submission date in the gdrive RCM
}
