% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm.R
\name{rcm_show}
\alias{rcm_show}
\title{display a row in the rcm based on the file id}
\usage{
rcm_show(rcm, file.id)
}
\arguments{
\item{rcm}{the RCM}

\item{file.id}{the file id, or a file id search term (see `?rcm_find_row_by_file.id` for details)}
}
\description{
display a row in the rcm based on the file id
}
