% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm.R
\name{rcm_sort_priority}
\alias{rcm_sort_priority}
\title{sort RCM by priority}
\usage{
rcm_sort_priority(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
the research cycle matrix sorted by 1. milestone passed, 2. days with HQ
}
\description{
sort RCM by priority
}
