% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_preprocessing.R
\name{rcm_standardised_columns}
\alias{rcm_standardised_columns}
\title{standardising columns}
\usage{
rcm_standardised_columns(rcm)
}
\description{
Regexing messy values into predefined categories and renames them. Applied by default in \link{\code{rcm_download}}
}
