% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_submission_expected}
\alias{rcm_submission_expected}
\title{Is an item expected for submission?}
\usage{
rcm_submission_expected(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
A subset of the research cycle matrix: all items with planned dates arrived/passed but don't have status HQ, validated, with field or with partner
}
\description{
Is an item expected for submission?
}
