% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_unit}
\alias{rcm_unit}
\title{which unit do RCM rows belong to?}
\usage{
rcm_unit(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
character vector  (with values "data", "reporting", "research design" or "GIS")
}
\description{
which unit do RCM rows belong to?
}
