% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_download.R
\name{rcm_update_from_subs}
\alias{rcm_update_from_subs}
\title{Update RCM status based on form validation submissions}
\usage{
rcm_update_from_subs(subs, rcm)
}
\arguments{
\item{subs}{the submissions as received from subs_download()}

\item{rcm}{the RCM as received by rcm_download()}
}
\description{
Update RCM status based on form validation submissions
}
