% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{status_with_field}
\alias{status_with_field}
\title{is with field?}
\usage{
status_with_field(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
is with field?
}
