% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{status_with_hq}
\alias{status_with_hq}
\title{is with HQ?}
\usage{
status_with_hq(rcm)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
logical vector
}
\description{
is with HQ?
}
