% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_download.R
\name{subs_download}
\alias{subs_download}
\title{downloads items submitted for validation through the shiny submission form (stored in gdrive)}
\usage{
subs_download()
}
\value{
a data frame with all submissions
}
\description{
downloads items submitted for validation through the shiny submission form (stored in gdrive)
}
