% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_download.R
\name{subs_rcm_rows}
\alias{subs_rcm_rows}
\title{find submission rows in the research cyclce matrix}
\usage{
subs_rcm_rows(subs, rcm)
}
\arguments{
\item{subs}{the submissions from subs_download()}

\item{rcm}{the research cycle matrix from rcm_download()}
}
\value{
a vector with row indices
}
\description{
find submission rows in the research cyclce matrix
}
\examples{
# get the matching rcm rows:
rcm<-rcm_download()
subs<-subs_download()
subs_in_rcm_indices<-subs_rcm_rows(subs,rcm)
# subset rcm
rcm[subs_in_rcm_indices,]
}
