% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_new_id_requests.R
\name{subs_with_new_id}
\alias{subs_with_new_id}
\title{Get only submissions where new ids were suggested}
\usage{
subs_with_new_id(subs)
}
\arguments{
\item{subs}{the submission list as received by subs_download()}
}
\value{
the subset of the submission list
}
\description{
Get only submissions where new ids were suggested
}
