% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo_create}
\alias{todo_create}
\title{make 'todo' item from rcm and subs}
\usage{
todo_create(rcm, subs, who = ".")
}
\arguments{
\item{rcm}{RCM from rcm_download(include_validated=T,include_archived=T)}

\item{subs}{Submissions from subs_download()}

\item{who}{your first name, i.e. "Karen". Allows regex (for example put "." to get all items)}
}
\value{
the todo item (silently)
}
\description{
make 'todo' item from rcm and subs
}
\seealso{
todo_download()
}
