% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo_delayed.R
\name{todo_delayed}
\alias{todo_delayed}
\title{which data unit items should have already been submitted but havent?}
\usage{
todo_delayed(rcm, days_since_planned_submission = 14)
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download()}

\item{days_since_planned_submission}{number of days grace period; if 14 (default), shows only items that are more than two weeks overdue}
}
\description{
which data unit items should have already been submitted but havent?
}
