% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo_delayed.R
\name{todo_delayed_to_csv}
\alias{todo_delayed_to_csv}
\title{create per country csv files listing all delayed items}
\usage{
todo_delayed_to_csv(todo, path = "./delayed/")
}
\arguments{
\item{todo}{see ?todo_download}
}
\description{
create per country csv files listing all delayed items
}
