% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo_download}
\alias{todo_download}
\title{Download 'todo' item list}
\usage{
todo_download(who = ".")
}
\arguments{
\item{who}{your first name, i.e. "Karen". Allows regex (for example put "." to get all items)}
}
\value{
the todo item (silently)
}
\description{
Download 'todo' item list
}
\details{
actually downloads the research cycle matrix and the submissions sheet and mushes them up into a todo list
The list is ordered by priority using 1. flagged as emergency? 2. Days
}
\seealso{
todo_download()
}
