% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo_next}
\alias{todo_next}
\title{show next on todo list}
\usage{
todo_next(todo, n = nrow(todo))
}
\arguments{
\item{todo}{'todo' list element (mix of rcm and subs; created with todo_download() )}

\item{n}{number of items to show}
}
\value{
the todo item (silently)
}
\description{
show next on todo list
}
\details{
will print the highest priority item last
}
