% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo_validate_next}
\alias{todo_validate_next}
\title{set the top item in the todo list to validated on the RCM}
\usage{
todo_validate_next(todo)
}
\arguments{
\item{todo}{'todo' list element (mix of rcm and subs; created with todo_download() )}
}
\value{
the todo item (silently), without the validated entry
}
\description{
set the top item in the todo list to validated on the RCM
}
\details{
will prompt for confirmation (type 'y' in console and hit enter; everything else will abort)
}
