% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{g_sheets_put}
\alias{g_sheets_put}
\title{wrapper for google sheets api}
\usage{
g_sheets_put(row, col, value,
  spreadsheetId = "1Quu2P6z-uA2H64eQENWJkNIOGIfnsdXgKNg4qdiCvXc", ...)
}
\description{
wrapper for google sheets api
}
