% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{rcm_change_value}
\alias{rcm_change_value}
\title{change a value on google drive based on the file.id}
\usage{
rcm_change_value(file.id, column, value)
}
\description{
change a value on google drive based on the file.id
}
