% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_gdrive.R
\name{rcm_comment}
\alias{rcm_comment}
\title{set an item's status to "validated" on google drive}
\usage{
rcm_comment(file.id, comment, overwrite = F)
}
\arguments{
\item{file.id}{the items file id name as a string}

\item{comment}{the comment as a single character string}

\item{overwrite}{if TRUE, will completely overwrite the current comment. Otherwise it will append the new comment to the existing one (with a date stamp)}
}
\description{
set an item's status to "validated" on google drive
}
