% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_status.R
\name{rcm_longest_with_hq}
\alias{rcm_longest_with_hq}
\title{is with field?}
\usage{
rcm_longest_with_hq(rcm, n = NULL, add.columns = c())
}
\arguments{
\item{rcm}{the research cycle matrix from rcm_download(raw=F)}
}
\value{
a subset of the RCM: only rows that are with HQ; only basic information columns. Sorted by added column "days.with.hq"
}
\description{
is with field?
}
