% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm_plots.R
\name{validation_timeline}
\alias{validation_timeline}
\title{plot validation timeline}
\usage{
validation_timeline(rcm)
}
\arguments{
\item{rcm}{the RCM as given by rcm_download()}
}
\description{
plot validation timeline
}
